requirements.txt
asgiref==3.6.0
Django==4.2
Pillow==9.5.0
sqlparse==0.4.3

django-embed-video==1.4.8